Move synth control element creation to Views file
